+++ src/pages/DatePickerPage.js
@@ -1,22 +1,51 @@
 import React, { useState } from 'react';
-import { TimePicker, Card, Select, Icon, Switch, Table, Collapse, CopyToClipboard, ThemeContext, Divider } from '../components';
+import { DatePicker, Card, Select, Icon, Table, Collapse, CopyToClipboard, ThemeContext, Divider } from '../components';
 import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
 import { tomorrow, coy } from 'react-syntax-highlighter/dist/esm/styles/prism';
+import { dateMask } from '../components/utils';
 
 const keys = ['property', 'description', 'default', 'type', 'value'];
 const rows = [
     { 
         property: 'value', 
-        description: 'Time value', 
+        description: 'Date value (handled by your logic)', 
+        default: '', 
+        type: 'string',
+        value: ''
+    },
+    { 
+        property: 'active', 
+        description: 'Date value (original Date value)', 
         default: '', 
         type: 'string',
         value: ''
     },
     {
-        property: 'onChange',
-        description: 'Invokes on time pick changes', 
+        property: 'onDate',
+        description: 'Invokes on date pick', 
         default: '', 
         type: 'function',
+        value: ''
+    },
+    { 
+        property: 'disabledDates',
+        description: 'Pass array of dates that not available for pick', 
+        default: '', 
+        type: 'Date[]',
+        value: ''
+    },
+    { 
+        property: 'minDate',
+        description: 'Set minimum endable date for pick', 
+        default: '', 
+        type: 'Date',
+        value: ''
+    },
+    { 
+        property: 'maxDate',
+        description: 'Set maximum endable date for pick', 
+        default: '', 
+        type: 'Date',
         value: ''
     },
     { 
@@ -38,21 +67,7 @@ const rows = [
         description: 'Set locale', 
         default: 'en', 
         type: 'string',
-        value: 'en | ru | kz'
-    },
-    { 
-        property: 'noSeconds', 
-        description: 'Leave only hours and minutes', 
-        default: 'false', 
-        type: 'boolean',
-        value: 'true | false'
-    },
-    { 
-        property: 'width', 
-        description: 'Set width', 
-        default: '', 
-        type: 'string | number',
-        value: ''
+        value: 'en | kz | ru | fr'
     },
     { 
         property: 'label', 
@@ -69,8 +84,15 @@ const rows = [
         value: ''
     },
     { 
-        property: 'required', 
-        description: 'Set required flag', 
+        property: 'hideWeekend',
+        description: 'If flag is set, weekend highlighting disabled', 
+        default: 'false', 
+        type: 'boolean',
+        value: 'true | false'
+    },
+    { 
+        property: 'hideCurrentDay',
+        description: 'If flag is set, today highlighting disabled', 
         default: 'false', 
         type: 'boolean',
         value: 'true | false'
@@ -90,11 +112,32 @@ const rows = [
         value: ''
     },
     { 
+        property: 'required', 
+        description: 'Set required flag', 
+        default: 'false', 
+        type: 'boolean',
+        value: 'true | false'
+    },
+    { 
         property: 'dark', 
         description: 'Set dark mode', 
         default: 'false', 
         type: 'boolean',
         value: 'true | false'
+    },
+    { 
+        property: 'cancelBtnName', 
+        description: 'Set cancel button name', 
+        default: 'Cancel', 
+        type: 'string',
+        value: ''
+    },
+    { 
+        property: 'dark', 
+        description: 'Set cancel button color', 
+        default: 'DatePicker color', 
+        type: 'string',
+        value: ''
     },
     { 
         property: 'className',
@@ -108,16 +151,16 @@ const rows = [
 const usage =
 `// Usage examples
 import React, { useState } from 'react';
-import { TimePicker, Select, Icon, Switch, Divider } from '@assenti/react-ui-components';
+import { DatePicker, Select, Icon, Divider } from '@assenti/react-ui-components';
 const colors = ['primary', 'info', 'success', 'error'];
 const sizes = ['default', 'medium', 'large'];
 const locales = ['en', 'kz', 'ru'];
 
 function Example() {
-    const [time, setTime] = useState('');
+    const [date, setDate] = useState('');
+    const [dateOriginal, setDateOriginal] = useState('');
     const [color, setColor] = useState('primary');
     const [size, setSize] = useState('default');
-    const [noSec, setNoSec] = useState(false);
     const [locale, setLocale] = useState('en');
 
     return (
@@ -154,26 +197,25 @@ function Example() {
                 value={locale}
                 onChange={v => setLocale(v)}/>
             <br/>
-            <Switch
-                check={noSec}
-                onChange={() => setNoSec(!noSec)}
-                color="primary"
-                className="my-10"
-                rightLabel="No seconds"
-                />
             <Divider/>
-            <TimePicker
+            <DatePicker
                 color={color}
                 size={size}
                 locale={locale}
-                noSeconds={noSec}
-                value={time}
+                placeholder="DD.MM.YYYY"
+                value={date}
                 clearable
                 dark={theme}
-                onClear={() => setTime('')}
-                width={200}
-                onChange={(value) => setTime(value)}
-                />
+                maxDate={new Date()}
+                onDate={(date) => {
+                    console.log(date)
+                    setDate(dateMask(date))
+                    setDateOriginal(date)
+                }}
+                active={dateOriginal}
+                onClear={() => setDate('')}
+                width={250}
+                onChange={(value) => setDate(value)}/>
         </>
     )
 }`
@@ -182,18 +224,18 @@ const colors = ['primary', 'info', 'success', 'error'];
 const sizes = ['default', 'medium', 'large'];
 const locales = ['en', 'kz', 'ru'];
 
-const TimePickerPage = () => {
-    const [time, setTime] = useState('');
+const DatePickerPage = () => {
+    const [date, setDate] = useState('');
+    const [dateOriginal, setDateOriginal] = useState('');
     const [color, setColor] = useState('primary');
     const [size, setSize] = useState('default');
-    const [noSec, setNoSec] = useState(false);
     const [locale, setLocale] = useState('en');
 
     return (
         <ThemeContext.Consumer>
             {theme => (
                 <div className="rui-page">
-                    <div className="rui-page-title">{'<TimePicker/>'} Component</div>
+                    <div className="rui-page-title">{'<DatePicker/>'} Component</div>
                     <Card dark={theme} header={<h4>Usage</h4>}>
                         <Select
                             items={colors}
@@ -227,26 +269,26 @@ const TimePickerPage = () => {
                             value={locale}
                             onChange={v => setLocale(v)}/>
                         <br/>
-                        <Switch
-                            check={noSec}
-                            onChange={() => setNoSec(!noSec)}
-                            color="primary"
-                            className="my-10"
-                            rightLabel="No seconds"
-                            />
                         <Divider/>
-                        <TimePicker
+                        <DatePicker
                             color={color}
                             size={size}
                             locale={locale}
-                            noSeconds={noSec}
-                            value={time}
+                            placeholder="DD.MM.YYYY"
+                            value={date}
                             clearable
+                            hideWeekend
                             dark={theme}
-                            onClear={() => setTime('')}
-                            width={200}
-                            onChange={(value) => setTime(value)}
-                            />
+                            maxDate={new Date()}
+                            onDate={(date) => {
+                                console.log(date)
+                                setDate(dateMask(date))
+                                setDateOriginal(date)
+                            }}
+                            active={dateOriginal}
+                            onClear={() => setDate('')}
+                            width={250}
+                            onChange={(value) => setDate(value)}/>
                         <Collapse 
                             icon="code" 
                             iconSize={18} 
@@ -277,4 +319,4 @@ const TimePickerPage = () => {
         </ThemeContext.Consumer>
     )
 }
-export default TimePickerPage;
+export default DatePickerPage;